feat: split navigation into marketplace, local models, inference, and openclaw
Implement Approach A (Thin Split) from Issue #6. This separates the
monolithic Hub into four focused destinations and simultaneously
resolves Issues #1–#5.

Changes:
- NEW: /local-models/ page for local model inventory (Ollama + GGUF)
- NEW: /openclaw/ page for OpenClaw instance management
- REFACTOR: /hub/ is now a pure inference center (runtime state only)
- REDIRECT: /hub/\ → /marketplace/\
- REMOVE: PullModelDialog from hub (download owned by marketplace only)
- FIX: Marketplace defaults to filter_library=gguf (Issue #1)
- FIX: Hub detail no longer uses fragile metadata path (Issue #2)
- FIX: Removed misplaced pull-model entry (Issue #3)
- FIX: Removed fake progress/cancel affordance (Issue #4)
- FIX: Hub detail no longer uses generic download path (Issue #5)
- UI: Sidebar now shows four entries: 模型市场, 本地模型管理,
  推理中心, OpenClaw 实例管理
- i18n: Added zh-CN and en translations for new nav items
